fix(polish): one owner for the snapshot rule, and a telemetry list that knows this generation (#2602)
Cloud review round 2 on #2603. One finding, plus a correction to my own previous
commit message.

THE CORRECTION FIRST. Commit 1b10945d claimed the repository had no OpenAI
date-snapshot handling and that the reviewer's premise did not check out. That
was wrong. `SettingsChangeTelemetry.stripDateSnapshotSuffix` has handled both
provider spellings since #158, and it is the exact rule I had just written a
second copy of. My sweep was scoped to `Sources/EnviousWisprLLM/` and the
settings view, so it returned a clean, complete-looking, empty answer — a wrong
SCOPE rather than a wrong pattern, which is the failure that does not announce
itself. The reviewer was right.

So the rule now has ONE owner. `ProviderModelID.withoutDateSnapshot` is the
provider-id fact both callers need, and both point at it. It did not stay on
`AIPolishModelClassifier` because that type's own note asks callers not to adopt
it elsewhere without revisiting placement, and this is that.

ONE DELIBERATE BEHAVIOUR CHANGE, stated because it touches telemetry. The shared
rule requires a real CALENDAR DATE, which the telemetry copy did not: an id whose
digits are not a date no longer collapses onto a base model, so `gpt-4-20259999`
reports `custom` rather than `gpt-4`. That is the honest answer for an id no
provider ships, and the suffix was already documented there as public and
non-sensitive, so nothing about the privacy boundary moves.

THE FINDING. `cloudModelAllowlist` is deny-by-default, so an unlisted model
reports `custom` and no unknown string leaks — but the signal is lost, and Luna
is the model this issue promotes to Recommended, so it would have become the
most-selected invisible model.

Enumerated from the live model lists of all three keys on 2026-09-02 rather than
from the existing entries, which is what the previous fix's own comment (#1770,
Gemini 3.x) shows this list needs: the same gap had reopened for every generation
shipped since. Added OpenAI 5.1 through 5.6 including all three codenames,
`gemini-3.8-flash`, `claude-fable-5-1` and `claude-opus-5`.

Full Debug lane: 7336 tests, verdict PASS, per-bundle sum reconciled. The
telemetry suites cover the changed normalisation and are green.
